refactor(web): extract VerticalBandSlider from EqualizerPanel
The interactive per-band gain slider was defined inline in the 490-line
EqualizerPanel file. Move VerticalBandSlider (and the BAND_STEP constant it
exclusively owns) into its own components/player/VerticalBandSlider.tsx. Pure
move, no behavior change; the panel keeps formatBandLabel since it builds the
slider's label prop. Drops the now-unused SliderPrimitive and EQ_MIN_DB/EQ_MAX_DB
imports from the panel.
